sparrow-data/api/requirements-fastapi.txt
pdf2image==1.16.2
datasets==2.10.1
Pillow==9.5.0
torch==1.13.1
torchvision
torchaudio
paddleocr==2.6.1.3
paddlepaddle==2.4.2
fastapi==0.95.2
python-multipart
uvicorn[standard]